requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
apispec==6.7.1
attrs==24.2.0
bech32==1.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cosmpy==0.9.2
ecdsa==0.19.0
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
grpcio==1.67.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
uagents==0.17.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
websockets==10.4
Werkzeug==3.1.3
yarl==1.17.1
